Some simple fixes for warnings added as part of slow-moving tidyup process.

+++ firtree/internal/lru-cache.h
@@ -168,7 +168,7 @@ class LRUCache
                     ageComp());
 
             int numToRemove = GetCacheSize() - m_IdealSize;
-            if(numToRemove > removableEntries.size())
+            if(numToRemove > (int)(removableEntries.size()))
             {
                 numToRemove = removableEntries.size();
             }
